C51 code is refactorized
